backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
pyjwt==2.8.0
werkzeug==3.0.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
